+++ app/src/main/java/org/simple/clinic/teleconsultlog/prescription/TeleconsultPrescriptionScreen.kt
@@ -1,48 +1,48 @@
 package org.simple.clinic.teleconsultlog.prescription
 
 import android.content.Context
-import android.os.Parcelable
-import android.util.AttributeSet
-import androidx.constraintlayout.widget.ConstraintLayout
+import android.view.LayoutInflater
+import android.view.ViewGroup
 import com.jakewharton.rxbinding3.appcompat.navigationClicks
 import com.jakewharton.rxbinding3.view.clicks
+import com.spotify.mobius.functions.Consumer
 import io.reactivex.Observable
-import io.reactivex.rxkotlin.ofType
+import io.reactivex.rxkotlin.cast
+import kotlinx.parcelize.Parcelize
 import org.simple.clinic.R
 import org.simple.clinic.ReportAnalyticsEvents
 import org.simple.clinic.databinding.ScreenTeleconsultPrescriptionBinding
 import org.simple.clinic.di.injector
-import org.simple.clinic.mobius.MobiusDelegate
 import org.simple.clinic.navigation.v2.Router
-import org.simple.clinic.navigation.v2.keyprovider.ScreenKeyProvider
+import org.simple.clinic.navigation.v2.ScreenKey
+import org.simple.clinic.navigation.v2.fragments.BaseScreen
 import org.simple.clinic.patient.Patient
 import org.simple.clinic.patient.displayLetterRes
 import org.simple.clinic.teleconsultlog.shareprescription.TeleconsultSharePrescriptionScreenKey
 import org.simple.clinic.util.UserClock
-import org.simple.clinic.util.unsafeLazy
 import org.simple.clinic.widgets.UiEvent
-import org.simple.clinic.widgets.hideKeyboard
 import java.util.UUID
 import javax.inject.Inject
 
-class TeleconsultPrescriptionScreen constructor(
-    context: Context,
-    attrs: AttributeSet?
-) : ConstraintLayout(context, attrs), TeleconsultPrescriptionUi, TeleconsultPrescriptionUiActions {
-
-  private var binding: ScreenTeleconsultPrescriptionBinding? = null
+class TeleconsultPrescriptionScreen : BaseScreen<
+    TeleconsultPrescriptionScreen.Key,
+    ScreenTeleconsultPrescriptionBinding,
+    TeleconsultPrescriptionModel,
+    TeleconsultPrescriptionEvent,
+    TeleconsultPrescriptionEffect,
+    Unit>(), TeleconsultPrescriptionUi, TeleconsultPrescriptionUiActions {
 
   private val toolbar
-    get() = binding!!.toolbar
+    get() = binding.toolbar
 
   private val teleconsultPrescriptionDoctorInfoView
-    get() = binding!!.teleconsultPrescriptionDoctorInfoView
+    get() = binding.teleconsultPrescriptionDoctorInfoView
 
   private val teleconsultPrescriptionMedicinesView
-    get() = binding!!.teleconsultPrescriptionMedicinesView
+    get() = binding.teleconsultPrescriptionMedicinesView
 
   private val nextButton
-    get() = binding!!.nextButton
+    get() = binding.nextButton
 
   @Inject
   lateinit var router: Router
@@ -53,70 +53,44 @@ class TeleconsultPrescriptionScreen constructor(
   @Inject
   lateinit var userClock: UserClock
 
-  @Inject
-  lateinit var screenKeyProvider: ScreenKeyProvider
-
-  private val screenKey by unsafeLazy {
-    screenKeyProvider.keyFor<TeleconsultPrescriptionScreenKey>(this)
-  }
-
-  private val events by unsafeLazy {
-    Observable
-        .merge(
-            backClicks(),
-            nextClicks()
-        )
-        .compose(ReportAnalyticsEvents())
-  }
-
-  private val delegate by unsafeLazy {
-    val uiRenderer = TeleconsultPrescriptionUiRenderer(this)
-
-    MobiusDelegate.forView(
-        events = events.ofType(),
-        defaultModel = TeleconsultPrescriptionModel.create(screenKey.teleconsultRecordId, screenKey.patientUuid),
-        init = TeleconsultPrescriptionInit(),
-        update = TeleconsultPrescriptionUpdate(),
-        effectHandler = effectHandler.create(this).build(),
-        modelUpdateListener = uiRenderer::render
-    )
-  }
-
-  override fun onAttachedToWindow() {
-    super.onAttachedToWindow()
-    delegate.start()
-  }
-
-  override fun onDetachedFromWindow() {
-    hideKeyboard()
-    delegate.stop()
-    binding = null
-    super.onDetachedFromWindow()
-  }
-
-  override fun onSaveInstanceState(): Parcelable {
-    return delegate.onSaveInstanceState(super.onSaveInstanceState())
-  }
-
-  override fun onRestoreInstanceState(state: Parcelable?) {
-    super.onRestoreInstanceState(delegate.onRestoreInstanceState(state))
-  }
-
-  override fun onFinishInflate() {
-    super.onFinishInflate()
-    if (isInEditMode) return
-
-    binding = ScreenTeleconsultPrescriptionBinding.bind(this)
+  override fun onAttach(context: Context) {
+    super.onAttach(context)
     context.injector<Injector>().inject(this)
   }
+
+  override fun defaultModel() = TeleconsultPrescriptionModel.create(
+      screenKey.teleconsultRecordId,
+      screenKey.patientUuid
+  )
+
+  override fun events() = Observable
+      .merge(
+          backClicks(),
+          nextClicks()
+      )
+      .compose(ReportAnalyticsEvents())
+      .cast<TeleconsultPrescriptionEvent>()
+
+  override fun createInit() = TeleconsultPrescriptionInit()
+
+  override fun createUpdate() = TeleconsultPrescriptionUpdate()
+
+  override fun createEffectHandler(viewEffectsConsumer: Consumer<Unit>) = effectHandler
+      .create(this)
+      .build()
+
+  override fun uiRenderer() = TeleconsultPrescriptionUiRenderer(this)
+
+  override fun bindView(layoutInflater: LayoutInflater, container: ViewGroup?) = ScreenTeleconsultPrescriptionBinding
+      .inflate(layoutInflater, container, false)
 
   override fun renderPatientDetails(patient: Patient) {
     val ageValue = patient.ageDetails.estimateAge(userClock)
     val patientGender = patient.gender
-    toolbar.title = context.getString(
+    toolbar.title = getString(
         R.string.screen_teleconsult_prescription_patient_details,
         patient.fullName,
-        context.getString(patientGender.displayLetterRes),
+        getString(patientGender.displayLetterRes),
         ageValue.toString()
     )
   }
@@ -155,6 +129,16 @@ class TeleconsultPrescriptionScreen constructor(
               medicalRegistrationId = medicalRegistrationId
           )
         }
+  }
+
+  @Parcelize
+  data class Key(
+      val patientUuid: UUID,
+      val teleconsultRecordId: UUID,
+      override val analyticsName: String = "Teleconsultation Prescription"
+  ) : ScreenKey() {
+
+    override fun instantiateFragment() = TeleconsultPrescriptionScreen()
   }
 
   interface Injector {